Compile expat even without entropy source when using configure
This is a temporary workaround for compilation errors when building the
library under old (or, rather, not very recent) Linux versions, it will
be soon made unnecessary by upgrading Expat to a later version fixing
this in the upstream sources.

This propagates 99be9a25f2f1d479371c80e85153b77961c3514f from wxWidgets
repository to the new submodule.