Consistently fix compilation problems for pre-C99 compilers
This reverts the parent commit and fixes all the problems due to
declaring variables not at the start of the block, which is not
supported by pre-C99 compilers such as MSVS 2008, by adding extra blocks
rather than separating the variable declarations and definitions as done
in the parent commit.

This should make it easier to merge with the upstream changes later and
avoid conflicts when we revert this commit (and the parent one) after
3.2.0. In case any conflicts still do occur, it should be clear enough
that the lines marked with "FIXME-C89" comment need to be removed when
resolving them.